Refactor v2 validator operators
